refactor: TextFiled를 공용으로 사용
