+++ frontend/src/components/ChannelCreateModal/ChannelCreateModal.stories.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import ChannelCreateModal from '@/components/ChannelCreateModal';
+import ChannelCreateModal from './ChannelCreateModal';
 
 storiesOf('ChannelCreateModal', module).add('basic', () => (
     <ChannelCreateModal />
