+++ frontend/src/components/Common/TextField/TextField.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import TextField from '@/components/TextField';
+import TextField from './TextField';
 
 storiesOf('TextField', module).add('basic', () => <TextField />);
